+++ GitTfsTest/Core/CommitParserTests.cs
@@ -16,7 +16,7 @@ namespace Sep.Git.Tfs.Test.Core
             {
                 return new[] {
                     new object[] { "git-tfs-id: foo;C123", true, 123 },
-                    new object[] { "git-tfs-id: handle more than Int32;C" + uint.MaxValue, true, uint.MaxValue },
+                    new object[] { "git-tfs-id: handle more than Int32;C" + int.MaxValue, true, int.MaxValue },
                     new object[] { "foo-tfs-id: bar;C123", false, 0 },
                     new object[] { "\ngit-tfs-id: foo;C234\n", true, 234 },
                     new object[] { "\r\ngit-tfs-id: foo;C345\r\n", true, 345 },
@@ -32,9 +32,9 @@ namespace Sep.Git.Tfs.Test.Core
 
         [Theory]
         [PropertyData("Cases")]
-        public void Run(string message, bool expectParsed, long expectId)
+        public void Run(string message, bool expectParsed, int expectId)
         {
-            long id;
+            int id;
             bool parsed = GitRepository.TryParseChangesetId(message, out id);
             Assert.Equal(expectParsed, parsed);
             if (parsed)
